refactor(server): Move RulesPolicy to zally-core #560
