+++ server/src/main/java/de/zalando/zally/rule/ApiValidator.kt
@@ -1,5 +1,7 @@
 package de.zalando.zally.rule
 
+import de.zalando.zally.core.RulesPolicy
+
 interface ApiValidator {
     fun validate(content: String, policy: RulesPolicy, authorization: String? = null): List<Result>
 }
