+++ server/src/main/java/de/zalando/zally/rule/SupportedRulesController.kt
@@ -1,5 +1,6 @@
 package de.zalando.zally.rule
 
+import de.zalando.zally.core.RulesPolicy
 import de.zalando.zally.dto.RuleDTO
 import de.zalando.zally.dto.RulesListDTO
 import de.zalando.zally.dto.SeverityBinder
